+++ src/modules/network/basic/src/half/half.cpp
@@ -97,7 +97,7 @@ static inline uint32_t _uint32_not( uint32_t a )
 // Negate
 static inline uint32_t _uint32_neg( uint32_t a )
 {
-  return (-a);
+  return (uint32_t)(-(int32_t)a);
 }
 
 // Extend sign
